+++ files/root/bin/my-startup.sh
@@ -3,22 +3,24 @@ cd /root/bin/
 
 if [ -n "$DOMAINS" ] ; then
 	DOMAINZ=(${DOMAINS})
-	# output is not success then exit
-	if test $(./le-run.sh | tee ../letsencrypt.init.log | grep -c 'Congratulations') -eq 0; then
-		echo "letsencrypt failed"
-		exit 1
+	MYDOMAIN=${DOMAINZ[0]}
+
+    # only generate domain if not exists
+	if [ ! -f /etc/letsencrypt/live/$MYDOMAIN/fullchain.pem ]; then
+		# output is not success then exit
+		if test $(./le-run.sh | tee ../letsencrypt.init.log | grep -c 'Congratulations') -eq 0; then
+			echo "letsencrypt failed"
+			exit 1
+		fi
+
+		echo "reloading domain ssl"
+
+		rm -f /app/etc/nginx/ssl/example-fullchain.pem
+		ln -s /etc/letsencrypt/live/$MYDOMAIN/fullchain.pem /app/etc/nginx/ssl/example-fullchain.pem
+		rm -f /app/etc/nginx/ssl/example-privkey.pem
+		ln -s /etc/letsencrypt/live/$MYDOMAIN/privkey.pem /app/etc/nginx/ssl/example-privkey.pem
+		service nginx reload
 	fi
-
-	echo "reloading domain ssl"
-	# linking the cert and restart nginx
-	# get the first folder name in live
-	DOMAIN=${DOMAINZ[0]}
-
-	rm -f /app/etc/nginx/ssl/example-fullchain.pem
-	ln -s /etc/letsencrypt/live/$DOMAIN/fullchain.pem /app/etc/nginx/ssl/example-fullchain.pem
-	rm -f /app/etc/nginx/ssl/example-privkey.pem
-	ln -s /etc/letsencrypt/live/$DOMAIN/privkey.pem /app/etc/nginx/ssl/example-privkey.pem
-	service nginx reload
 fi
 
 # hitcounter instruction: link nginx conf, provide AWS ENV, then setup contab
